+++ server/tests/backends/sql_translator/test_snowflake_translator_steps.py
@@ -69,10 +69,10 @@ def sql_retrieve_city(t):
     return t
 
 
-def snowflake_query_describer(t) -> Union[Dict[str, str], None]:
+def snowflake_query_describer(domain, query_string=None) -> Union[Dict[str, str], None]:
     connection = get_connection()
     with connection.cursor() as cursor:
-        describe_res = cursor.describe(t)
+        describe_res = cursor.describe(domain if domain else query_string)
         res = {r.name: type_code_mapping.get(r.type_code) for r in describe_res}
         return res
 
@@ -93,6 +93,10 @@ def test_sql_translator_pipeline(case_id, case_spec_file_path, get_engine):
     data_to_insert.to_sql(
         name=case_id.replace('/', ''), con=get_engine, index=False, if_exists='replace', chunksize=1
     )
+    for input in spec['other_inputs']:
+        pd.read_json(json.dumps(spec['other_inputs'][input]), orient='table').to_sql(
+            name=input, con=get_engine, index=False, if_exists='replace', chunksize=1
+        )
 
     steps = spec['step']['pipeline']
     steps.insert(0, {'name': 'domain', 'domain': f'SELECT * FROM {case_id.replace("/", "")}'})
@@ -109,7 +113,9 @@ def test_sql_translator_pipeline(case_id, case_spec_file_path, get_engine):
     result: pd.DataFrame = execute(get_connection(), query)
 
     # Drop created table
-    execute(get_connection(), f'DROP TABLE {case_id.replace("/", "")}', False)
+    execute(get_connection(), f'DROP TABLE IF EXISTS {case_id.replace("/", "")}', False)
+    for input in spec['other_inputs']:
+        execute(get_connection(), f'DROP TABLE IF EXISTS {input}', False)
 
     # Compare result and expected (from fixture file)
     pandas_result_expected = pd.read_json(json.dumps(spec['expected']), orient='table')