+++ tests/unit/translator.spec.ts
@@ -80,6 +80,7 @@ describe('translator registration', () => {
       'mongo42',
       'pandas',
       'pandas-no_joins',
+      'snowflake',
     ]);
     expect(backendsSupporting('domain')).toEqual([
       'dummy',
@@ -89,9 +90,22 @@ describe('translator registration', () => {
       'mongo42',
       'pandas',
       'pandas-no_joins',
+      'snowflake',
     ]);
-    expect(backendsSupporting('append')).toEqual(['mongo36', 'mongo40', 'mongo42', 'pandas']);
-    expect(backendsSupporting('join')).toEqual(['mongo36', 'mongo40', 'mongo42', 'pandas']);
+    expect(backendsSupporting('append')).toEqual([
+      'mongo36',
+      'mongo40',
+      'mongo42',
+      'pandas',
+      'snowflake',
+    ]);
+    expect(backendsSupporting('join')).toEqual([
+      'mongo36',
+      'mongo40',
+      'mongo42',
+      'pandas',
+      'snowflake',
+    ]);
   });
 
   it('should throw an error if backend is not available', () => {
@@ -109,6 +123,7 @@ describe('translator registration', () => {
       'mongo42',
       'pandas',
       'pandas-no_joins',
+      'snowflake',
     ]);
   });
 });