+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## Unreleased
+
+
+### [0.57.0] - 2021-08-26
+
+### Added
+
+- Mongo translator: `notnull` and `isnull` condition operator now always compare to literal null.
+- Vqb: Added UnuiqueGroup Step to snowflake
+
+## [0.56.0] - 2021-08-24
+
+### Added
+
+- New step: trim (for mongo >4 and pandas) by @CharlesRngrd
+- Translators: new snowflake translator
+
 ## [0.55.0] - 2021-08-09
 
 ### Added
@@ -830,6 +847,10 @@
 ### Added
 
 - Initial version, showtime!
+
+
+[0.56.0]: https://github.com/ToucanToco/weaverbird/compare/v0.55.0...v0.56.0
+[0.55.0]: https://github.com/ToucanToco/weaverbird/compare/v0.54.0...v0.55.0
 [0.54.0]: https://github.com/ToucanToco/weaverbird/compare/v0.53.1...v0.54.0
 [0.53.1]: https://github.com/ToucanToco/weaverbird/compare/v0.53.0...v0.53.1
 [0.53.0]: https://github.com/ToucanToco/weaverbird/compare/v0.52.0...v0.53.0